mt5-fastapi/requirements.txt
numpy<2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
MetaTrader5==5.0.47
pydantic==2.5.0
python-dotenv==1.0.0
